+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/domain/auth/Authority.java
@@ -1,8 +1,10 @@
 package gov.healthit.chpl.domain.auth;
 
-public class Authority {
+public final class Authority {
+    private Authority() {}
     public static final String ROLE_ADMIN = "ROLE_ADMIN";
     public static final String ROLE_ONC = "ROLE_ONC";
+    public static final String ROLE_ONC_STAFF = "ROLE_ONC_STAFF";
     public static final String ROLE_USER_CREATOR = "ROLE_USER_CREATOR";
     public static final String ROLE_ACB = "ROLE_ACB";
     public static final String ROLE_ATL = "ROLE_ATL";