+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/logging/Loggable.java
@@ -1,0 +1,16 @@
+package gov.healthit.chpl.logging;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+import org.springframework.stereotype.Component;
+
+@Component
+@Target({ElementType.TYPE, ElementType.METHOD})
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Loggable {
+    boolean params() default false;
+    boolean result() default false;
+}