+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/manager/InvitationManager.java
@@ -96,6 +96,20 @@ public class InvitationManager extends SecuredManager {
 
     @Transactional
     @PreAuthorize("@permissions.hasAccess(T(gov.healthit.chpl.permissions.Permissions).INVITATION, "
+            + "T(gov.healthit.chpl.permissions.domains.InvitationDomainPermissions).INVITE_ONC_STAFF)")
+    public InvitationDTO inviteOncStaff(String emailAddress)
+            throws UserCreationException, UserRetrievalException, UserPermissionRetrievalException {
+        InvitationDTO dto = new InvitationDTO();
+        dto.setEmail(emailAddress);
+        dto.setPermission(userPermissionDao.getPermissionFromAuthority(Authority.ROLE_ONC_STAFF));
+        Date now = new Date();
+        dto.setInviteToken(Util.md5(emailAddress + now.getTime()));
+
+        return createInvitation(dto);
+    }
+
+    @Transactional
+    @PreAuthorize("@permissions.hasAccess(T(gov.healthit.chpl.permissions.Permissions).INVITATION, "
             + "T(gov.healthit.chpl.permissions.domains.InvitationDomainPermissions).INVITE_CMS)")
     public InvitationDTO inviteCms(String emailAddress)
             throws UserCreationException, UserRetrievalException, UserPermissionRetrievalException {
@@ -188,26 +202,16 @@ public class InvitationManager extends SecuredManager {
         Authentication authenticator = AuthUtil.getInvitedUserAuthenticator(invitation.getLastModifiedUserId());
         SecurityContextHolder.getContext().setAuthentication(authenticator);
 
-        UserDTO existingUser = null;
         try {
-            existingUser = userManager.getByNameOrEmail(user.getSubjectName());
+            UserDTO existingUser = userManager.getByNameOrEmail(user.getEmail());
+            if (existingUser != null) {
+                throw new InvalidArgumentsException(msgUtil.getMessage("user.accountAlreadyExists", user.getEmail()));
+            }
+        } catch (UserRetrievalException urex) {
+            //ignore
         } catch (MultipleUserAccountsException ex) {
             //hitting this block means there are multiple users registered with this account.
             //don't let them make a new one!
-            throw new InvalidArgumentsException(msgUtil.getMessage("user.accountAlreadyExists", user.getSubjectName()));
-        }
-        if (existingUser != null) {
-            throw new InvalidArgumentsException(msgUtil.getMessage("user.accountAlreadyExists", user.getSubjectName()));
-        } else {
-            try {
-                existingUser = userManager.getByNameOrEmail(user.getEmail());
-            } catch (MultipleUserAccountsException ex) {
-                //hitting this block means there are multiple users registered with this account.
-                //don't let them make a new one!
-                throw new InvalidArgumentsException(msgUtil.getMessage("user.accountAlreadyExists", user.getEmail()));
-            }
-        }
-        if (existingUser != null) {
             throw new InvalidArgumentsException(msgUtil.getMessage("user.accountAlreadyExists", user.getEmail()));
         }
 
@@ -230,7 +234,7 @@ public class InvitationManager extends SecuredManager {
     }
 
     @Transactional
-    public UserDTO confirmAccountEmail(InvitationDTO invitation) throws UserRetrievalException {
+    public UserDTO confirmAccountEmail(InvitationDTO invitation) throws UserRetrievalException, MultipleUserAccountsException {
         Authentication authenticator = AuthUtil.getInvitedUserAuthenticator(invitation.getLastModifiedUserId());
         SecurityContextHolder.getContext().setAuthentication(authenticator);
 
@@ -248,7 +252,7 @@ public class InvitationManager extends SecuredManager {
             userDao.update(user);
             invitationDao.delete(invitation.getId());
 
-            String activityDescription = "User " + user.getSubjectName() + " was confirmed.";
+            String activityDescription = "User " + user.getEmail() + " was confirmed.";
             try {
                 activityManager.addActivity(ActivityConcept.USER, user.getId(), activityDescription, origUser, user,
                         user.getId());
@@ -347,7 +351,6 @@ public class InvitationManager extends SecuredManager {
     private UserDTO constructUser(InvitationDTO invitation, CreateUserRequest user) {
         UserDTO userDto = new UserDTO();
         userDto.setTitle(user.getTitle());
-        userDto.setSubjectName(user.getSubjectName());
         userDto.setPermission(invitation.getPermission());
         userDto.setPhoneNumber(user.getPhoneNumber());
         userDto.setPasswordResetRequired(false);