+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/listing/Edition2015PendingListingValidator.java
@@ -26,6 +26,7 @@ import gov.healthit.chpl.validation.pendingListing.reviewer.edition2015.Attested
 import gov.healthit.chpl.validation.pendingListing.reviewer.edition2015.CqmAttestedCriteriaReviewer;
 import gov.healthit.chpl.validation.pendingListing.reviewer.edition2015.GapAllowedReviewer;
 import gov.healthit.chpl.validation.pendingListing.reviewer.edition2015.InvalidCriteriaCombinationReviewer;
+import gov.healthit.chpl.validation.pendingListing.reviewer.edition2015.MeasureValidityReviewer;
 import gov.healthit.chpl.validation.pendingListing.reviewer.edition2015.OldCriteriaWithoutIcsReviewer;
 import gov.healthit.chpl.validation.pendingListing.reviewer.edition2015.PrivacyAndSecurityCriteriaReviewer;
 import gov.healthit.chpl.validation.pendingListing.reviewer.edition2015.RequiredData2015Reviewer;
@@ -142,6 +143,10 @@ public class Edition2015PendingListingValidator extends PendingValidator {
     @Qualifier("pendingGapAllowedReviewer")
     private GapAllowedReviewer gapAllowedReviewer;
 
+    @Autowired
+    @Qualifier("pendingMeasureValidityReviewer")
+    private MeasureValidityReviewer measureReviewer;
+
     private List<Reviewer> reviewers;
 
     @Override
@@ -173,6 +178,7 @@ public class Edition2015PendingListingValidator extends PendingValidator {
             reviewers.add(testFunctionalityAllowedByRoleReviewer);
             reviewers.add(testStandardReviewer);
             reviewers.add(gapAllowedReviewer);
+            reviewers.add(measureReviewer);
         }
         return reviewers;
     }