+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/validation/listing/package-info.java
@@ -1,4 +1,0 @@
-/**
- * Code to validate a variety of listing editions and types.
- */
-package gov.healthit.chpl.validation.listing;